ingestion/requirements.txt
fastcrc==0.3.5
psycopg2==2.9.11
pyserial==3.5
